fixes and response messages added

+++ api/controllers/pizzaController.js
@@ -4,6 +4,8 @@
 // Pizza = mongoose.model('Pizza');
 
 require("underscore");
+const path = require('path');
+
 var Pizza = require('../models/Pizza');
 var Topping = require('../models/Topping');
 var Order = require('../models/Order');
@@ -12,78 +14,57 @@ var OrderItem = require('../models/OrderItem');
 var connection = require('../../db');
 
 
-console.log("in controller");
+const PRICE_STANDARD = 5.00;
+const PRICE_LARGE = 8.50;
+
+// console.log("in controller");
 function response(data) {
   dataObj = JSON.parse(data.body);
   let resp;
   if (dataObj.error) {
-    // resp.
   }
 }
 
 exports.list_all_pizza = function (req, res) {
 
-  // console.log("in list all pizzA");
+  var requestParams = req.params;
+  if (requestParams) {
+    var arr = [];
+    Pizza.findAll({ attributes: ['id'], raw: true }).then(result => {
+      result.forEach(element => {
+        arr.push(element.id);
+      });
 
-  // Pizza.find({}, function (err, pizza) {
-  //   if (err)
-  //     res.send(err);
-  //   res.json(pizza);
-  // });
+      if (arr.length > 0)
+        res.status(200).send(arr);
+      else
+        res.status(404).send("No pizza exists.");
 
-  // connection.query('SELECT * FROM Pizza', function(err, results) {
-
-  //   res.send(results);
-  // });
-  Pizza.findAll({ raw: true }).then(result => {
-    console.log(result);
-  });
-
-
+    });
+  } else {
+    res.status(404).send("No pizza exists.");
+  }
 };
 
 exports.create_pizza = function (req, res) {
 
   var request = req.body;
+
+  if (request['id'])
+    delete request['id'];
+
   Pizza.create(request)
     .then(function (newModel) {
+
       console.log(newModel.dataValues);
+      if (newModel.dataValues) {
+        var uri = req.protocol + '://' + req.get('host') + req.originalUrl;
+        res.setHeader('Location', uri + "/" + newModel.dataValues.id);
+        res.status(201).send("Created new pizza.");
+      } else {
+        res.status(400).send("Invalid input.")
+      }
     });
-
-
-
-
-  // console.log(req.body);
-
-  // var sql = "INSERT INTO Pizza (name, size,price) VALUES ('"+req.body.name+"', '"+req.body.size+"','"+req.body.price+"')";
-  // connection.query(sql, function (err, result) {
-  //   if (err)
-  //       res.send(err);
-  //     res.json("1 record inserted");
-  //   console.log("1 record inserted");
-  // });
-
-  // var new_pizza = new Pizza(req.body);
-  // new_pizza.save(function (err, pizza) {
-  //   if (err)
-  //     res.send(err);
-  //   res.json(pizza);
-  // });
-  //   Pizza.nextCount(function(err,count){
-  //    if(count){
-  //     req.body["id"]=count;
-  //     var new_pizza = new Pizza(req.body);
-  //     console.log(new_pizza);
-  //     new_pizza.save(function (err, pizza) {console.log("here");
-  //       if (err)
-  //         res.send(err);
-  //       res.json(pizza);
-  //     });
-  //    }
-  //  });
-
-  //  console.log(req.body);
-
 };
 
 exports.get_pizza = function (req, res) {
@@ -91,19 +72,18 @@ exports.get_pizza = function (req, res) {
   var request = req.params;
   console.log("in get pizza");
 
-
   Pizza.findOne({
     where: {
       id: request.id
     }, raw: true
   }).then(result => {
-    console.log(result);
+    if (result) {
+      res.status(201).send(result);
+    } else {
+      res.status(404).send("Pizza not found.");
+    }
+    // console.log(result);
   });
-  // Pizza.find({ 'id': request.id }, function (err, pizza) {
-  //   if (err)
-  //     res.send(err);
-  //   res.json(pizza);
-  // });
 };
 
 
@@ -122,40 +102,38 @@ exports.update_pizza = function (req, res) {
       if (pizza) {
         pizza.updateAttributes(requestBody)
           .then(function (newPizza) {
-            console.log(newPizza.dataValues);
-          })
+            res.status(404).send("Pizza not found.");
+            // console.log(newPizza.dataValues);
+            res.status(204).send("Update okay.");
+
+          }).catch(function (err) {
+            res.status(400).send("Invalid pizza supplied.");
+          });
+      } else {
+        res.status(404).send("Pizza not found.");
       }
-    })
 
-
-
-  // Pizza.findOneAndUpdate({ 'id': req.params.id }, req.body, { new: true }, function (err, pizza) {
-  //   // console.log(res);
-  //   console.log(err);
-  //   if (err)
-  //     pizza.send(err);
-  //   res.json(pizza);
-
-  //   // return ({"message":"successful"});
-  // });
+    });
 };
 
 exports.delete_pizza = function (req, res) {
 
   var requestBody = req.params.id;
-  Pizza.destroy({
-    where: {
-      id: req.params.id
-    }
-  }).then(function (result) {
-    console.log(result);
-
-  });
-  // Pizza.findOneAndRemove({ 'id': req.params.id }, function (err, pizza) {
-  //   if (err)
-  //     res.send(err);
-  //   res.json(pizza);
-  // });
+  if (requestBody) {
+    Pizza.destroy({
+      where: {
+        id: req.params.id
+      }
+    }).then(function (result) {
+      if (result) {
+        res.status(204).send("Deleted.");
+      } else {
+        res.status(404).send("Pizza not found.");
+      }
+    });
+  } else {
+    res.status(404).send("Invalid Id supplied.");
+  }
 };
 
 
@@ -172,45 +150,51 @@ exports.create_topping = function (req, res) {
   }).then(result => {
     // console.log(result);
     if (result != null) {
+
       if (requestBody['id'])
         delete requestBody['id'];
+
       requestBody['pizzaId'] = parseInt(requestParams.pizzaId);
 
       Topping.create(requestBody)
         .then(function (newModel) {
-          console.log(newModel.dataValues);
+          if (newModel.dataValues) {
+            var uri = req.protocol + '://' + req.get('host') + req.originalUrl;
+            res.setHeader('Location', uri + "/" + newModel.dataValues.id);
+            res.status(201).send("Created new Topping for pizza.");
+          } else {
+
+            res.status(404).send("No toppings found.");
+          }
+          // console.log(newModel.dataValues);
         });
     }
 
-
+  }).catch(function (err) {
+    res.status(400).send("Invalid input.");
   });
-  // Pizza.find({ 'id': req.params.pizzaId }, function (err, pizza) {
-  //   if (err) {
-  //     res.send(err);
-  //   } else {
-  //     req.body["pizza_id"] = Number(req.params.pizzaId);
-  //     var newTopping = new Topping(req.body);
-  //     newTopping.save(function (err, topping) {
-  //       if (err)
-  //         res.send(err);
-  //       res.json(topping);
-  //     });
-  //   }
-
-  // });
 };
 
 
 exports.get_all_toppings = function (req, res) {
-  // Topping.find({ 'pizza_id': req.params.pizzaId }, function (err, pizza) {
-  //   if (err)
-  //     res.send(err);
-  //   res.json(pizza);
-  // });
+
   var requestParams = req.params;
-  Topping.findAll({ where: { pizzaId: requestParams.pizzaId }, raw: true }).then(result => {
-    console.log(result);
-  });
+  if (requestParams) {
+    var arr = [];
+    Topping.findAll({ where: { pizzaId: requestParams.pizzaId }, attributes: ['id'], raw: true }).then(result => {
+      result.forEach(element => {
+        arr.push(element.id);
+      });
+
+      if (arr.length > 0)
+        res.status(200).send(arr);
+      else
+        res.status(400).send("Toppings not found.");
+
+    });
+  } else {
+    res.status(404).send("Specified pizza id not found.");
+  }
 };
 
 
@@ -223,15 +207,33 @@ exports.get_topping = function (req, res) {
 
   var requestParams = req.params;
   Topping.findOne({ where: { 'pizzaId': req.params.pizzaId, 'id': req.params.toppingId }, raw: true }).then(result => {
-    console.log(result);
+    if (result) {
+      res.status(201).send(result);
+    } else {
+      res.status(404).send("Pizza or Topping not found.");
+    }
+  }).catch(function (err) {
+    res.status(400).send("Invalid ID(s) supplied..");
   });
+
 };
 
 
 exports.delete_topping = function (req, res) {
+
   Topping.destroy({ where: { 'pizza_id': req.params.pizzaId, 'id': req.params.toppingId }, raw: true }.then(result => {
     console.log(result);
-  }));
+
+    if (result) {
+      res.status(204).send("Deleted.");
+    } else {
+      res.status(404).send("Pizza not found.");
+    }
+  })).catch(function (err) {
+    res.status(400).send("Invalid IDs.");
+  });
+
+
 };
 
 
@@ -260,45 +262,65 @@ exports.create_order = function (req, res) {
 
   Order.create(orderData)
     .then(function (newOrder) {
-      console.log(newOrder.dataValues.id);
+      // console.log(newOrder.dataValues.id);
       oId = newOrder.dataValues.id;
 
       req.body["orderItems"].forEach(element => {
         element["orderId"] = oId;
+
+        Pizza.findOne({ where: { id: element["pizzaId"] }, attributes: ['size', 'price'], raw: true }).then(function (pizza) {
+          // console.log(pizza);
+          var pizzaValue = 0;
+          if (pizza.size == 'Large')
+            pizzaValue = PRICE_LARGE + pizza.price;
+          else
+            pizzaValue = PRICE_STANDARD + pizza.price;
+
+
+          totalCost = totalCost + (pizzaValue * element["quantity"]);
+          console.log(totalCost);
+        });
+
         OrderItem.create(element)
           .then(function (newModel) {
-
-            console.log(newModel.dataValues);
-
-            // if(newModel.dataValues.pizzaId)
-            // {
-            //   pId=newModel.dataValues.pizzaId;
-            //   var orderData = {
-            //     "totalprice":request["totalPrice"],
-            //     "recipient":request["recipient"],
-            //     "orderItemId":newModel.dataValues.id,
-            //   };
-
-
-
             // console.log(newModel.dataValues);
-            // console.log(newModel.dataValues.pizzaId);
-            // console.log(newModel.dataValues.quantity);
+
+
+            orderData["id"] = oId;
+            orderData["totalPrice"] = totalCost;
+
+            Order.findOne({ where: { id: oId } })
+              .then(function (order) {
+                console.log(order.dataValues);
+                // Check if record exists in db
+                if (order) {
+                  console.log('in update');
+                  order.updateAttributes(orderData)
+                    .then(function (newOrder) {
+                      console.log(newOrder.dataValues);
+
+                    }).catch(function (err) {
+                      console.log(err);
+                    });
+                } else {
+                }
+
+              });
+
+
 
 
           });
-        // var orderItem = new OrderItem(element);
-        // console.log(orderItem);
-
-        // orderItem.save(function (err, item) {
-        //   console.log(item);
-        //   if (err)
-        //     res.send(err);
-
-        //   res.json(item);
-        // });
-
       });
+
+
+      var uri = req.protocol + '://' + req.get('host') + req.originalUrl;
+      res.setHeader('Location', uri + "/" + newOrder.dataValues.id);
+      res.status(201).send("Order successfully created.");
+
+    })
+    .catch(function (err) {
+      res.status(400).send("Invalid Order.");
     });
   // }
   // Pizza.find({'id':req}, function(err, pizza) {
@@ -314,10 +336,17 @@ exports.get_all_orders = function (req, res) {
   if (requestParams) {
     var arr = [];
     Order.findAll({ attributes: ['id'], raw: true }).then(result => {
-      result.forEach(element => {
-        arr.push(element.id);
-      });
-      console.log(arr);
+      if (result.length > 0) {
+        result.forEach(element => {
+          arr.push(element.id);
+        });
+        res.status(200).send(arr);
+        // console.log(arr);
+      } else {
+        res.status(404).send("No orders found.");
+      }
+    }).catch(function (err) {
+      res.status(404).send("Invalid ID.");
     });
   }
 
@@ -329,37 +358,47 @@ exports.get_order = function (req, res) {
   if (requestParams) {
 
     Order.findOne({ where: { id: requestParams.orderId }, raw: true }).then(result => {
-      console.log(result);
+      // console.log(result);
+      if (result) {
+        OrderItem.findAll({ where: { orderId: requestParams.orderId }, raw: true }).then(items => {
+          // console.log(items);
 
-      OrderItem.findAll({ where: { orderId: requestParams.orderId }, raw: true }).then(items => {
-        console.log(items);
-  
-        
-        result["orderItems"] = items;
-        console.log(result);
-      });
+          result["orderItems"] = items;
 
+          // console.log(result);
+        });
 
+        res.status(200).send(result);
+      } else {
+        res.status(400).send("Order not found");
+      }
+
+    }).catch(function (err) {
+      res.status(400).send("Invalid ID supplied");
     });
   }
 };
 
 
-exports.delete_order= function(req, res) {
+exports.delete_order = function (req, res) {
   var requestBody = req.params.id;
   OrderItem.destroy({
     where: {
       orderId: req.params.orderId
     }
   }).then(function (resultItems) {
-    console.log(resultItems);
+    // console.log(resultItems);
     Order.destroy({
       where: {
         id: req.params.orderId
       }
     }).then(function (result) {
-      console.log(result);
-  
+      // console.log(result);
+      res.status(204).send("Deletion successful");
+    }).catch(function (err) {
+      res.status(404).send("Order not found.");
     });
+  }).catch(function (err) {
+    res.status(400).send("Invalid ID supplied");
   });
 };
